src/main/java/t1/ismailov/SpringSecurityJWT/service/AuthService.java
package t1.ismailov.SpringSecurityJWT.service;

import t1.ismailov.SpringSecurityJWT.dto.AuthRequest;
import t1.ismailov.SpringSecurityJWT.dto.AuthResponse;
import t1.ismailov.SpringSecurityJWT.dto.RegistrationUserDto;
import t1.ismailov.SpringSecurityJWT.exception.AppError;
import t1.ismailov.SpringSecurityJWT.security.jwt.JwtTokenUtils;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.BadCredentialsException;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.stereotype.Service;
import org.springframework.web.bind.annotation.RequestBody;


@Service
@RequiredArgsConstructor
public class AuthService {

    private final UserService userService;
    private final JwtTokenUtils jwtTokenUtils;
    private final AuthenticationManager authenticationManager;


    public ResponseEntity<?> createAuthToken(@RequestBody AuthRequest authRequest) {
        try {
            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authRequest.getUsername(),
                    authRequest.getPassword()));
            var token = jwtTokenUtils.generateToken(authRequest.getUsername());
            return ResponseEntity.ok(new AuthResponse(token));
        } catch (BadCredentialsException e) {
            return new ResponseEntity<>(new AppError(HttpStatus.UNAUTHORIZED.value(),
                    "Неправильный логин или пароль"), HttpStatus.UNAUTHORIZED);
        }

    }


    public ResponseEntity<?> createUser(@RequestBody RegistrationUserDto regUserDto) {
        if (!regUserDto.getPassword().equals(regUserDto.getConfirmPassword())) {
            return new ResponseEntity<>(new AppError(HttpStatus.BAD_REQUEST.value(), "Пароли не совпадают"),
                    HttpStatus.BAD_REQUEST);
        }
        if (userService.findByUsername(regUserDto.getUsername()).isPresent()) {
            return new ResponseEntity<>(new AppError(HttpStatus.BAD_REQUEST.value(), "Пользователь уже существует"),
                    HttpStatus.BAD_REQUEST);
        }

        userService.createUser(regUserDto);

        return ResponseEntity.ok().body(String.format("Пользователь %s создан", regUserDto.getUsername()));
    }

}
